Phase 10 prep: sweep GCC 15 / libstdc++ 15 fallout
Resolute (26.04) ships GCC 15 / libstdc++ 15 / Boost 1.90. GCC 15's
C++20-default-mode upgrades catch a pile of older patterns that
compiled cleanly under GCC 13, and the C++23 flip later in this
series escalates more of them to -Werror. Land all the
base-independent fixes first so Noble stays green and the FROM bump
itself stays focused on the LTS-specific changes.

Drop the template-id from constructors/destructors that spelled
themselves with the class template parameter (e.g. `atomic<T>()`):
the template-id-cdtor form is ill-formed in C++20. Touched
drizzled/atomics.h, drizzled/constrained_value.h, and
unittests/pthread_atomics_test.cc.

Replace `volatile`-qualified ++ / -- with explicit read-modify-write
(`x = x + 1`). C++20 deprecated the operator forms (P1152); the
explicit form has identical semantics and works under every
standard. Touched plugin/innobase/include/sync0rw.ic
(single-threaded direct paths, already documented as such in the
function headers), drizzled/atomic/pthread_traits.h
(mutex-protected fetch_and_* helpers), and
drizzled/locking/global.cc (reader/writer counters already accessed
under LOCK_global_read_lock). Also split the chained
`abort_loop = select_thread_in_use = false;` in
drizzled/drizzled.cc that tripped the same deprecation as a
value-of-assignment use.

drizzled/create_field.h's default constructor only initialised
`after`; GCC 15's tighter escape analysis flags every other public
member as may-be-used-uninitialised when CreateField() is fed
through Item_sum_distinct::setup(). Convert to in-class member
initialisers and `= default` the constructor.

client/drizzle.cc had a pre-existing latent bug: a bitwise & between
drizzle_column_type() (a drizzle_column_type_t) and
DRIZZLE_COLUMN_FLAGS_UNSIGNED (a drizzle_column_flags_t) that
happened to be non-zero for every TINY column. C++20 deprecates
mixed-enum bitwise ops; the static_cast preserves the existing
pre-bug behaviour, which main.drizzle_client's .result is keyed to.
Worth revisiting once we're willing to update the .result.

Migrate the deprecated `<boost/detail/atomic_count.hpp>` include in
drizzled/drizzled.{h,cc} to `<boost/smart_ptr/detail/atomic_count.hpp>`;
the new spelling has been in Boost since well before our floor.

Verification on Noble pending in the layering pass; each fix is a
pure no-op rewrite or modernisation that does not depend on the
new toolchain.

+++ client/drizzle.cc
@@ -3211,10 +3211,11 @@ print_table_data(drizzle_result_st *result)
       // Room for "NULL"
       length=4;
     }
-    if ((length < 5) and 
+    if ((length < 5) and
       (server_type == ServerDetect::SERVER_DRIZZLE_FOUND) and
       (drizzle_column_type(field) == DRIZZLE_COLUMN_TYPE_TINY) and
-      (drizzle_column_type(field) & DRIZZLE_COLUMN_FLAGS_UNSIGNED))
+      (static_cast<int>(drizzle_column_type(field))
+       & static_cast<int>(DRIZZLE_COLUMN_FLAGS_UNSIGNED)))
     {
       // Room for "FALSE"
       length= 5;
